this branch is the failed branch
so here is the gist of what happened. two major issues
1. the recipe model is being passed to the sub-compenent correctly however for some reason it is completely udnefined thus all you will see is [object object] in the cards list. I have no idea why this is the case there are no errors thrown in the console and no deviations from the example code on screen was noticed after several watches.
2. when implementing the add ingredients function in the edit shopping section of the project the method shorts out and transfers the view back over to the default component and away from the shopping list section. close observation reveals that a new list card is created but appears to be completely empty. to me this indicates yet another example of an undefined custom object being passed and yet becomes undefined for some reason. again there are no errors thrown.  prior to these changes theere was no issue viewing the entire project.

as a possible attempt to correct this I will compare the commits on this branch to the commits created on branch 84-89 max final example which will be loaded with the instructor's example code. it is my hope that that line by line analysis could reveal some kind of oversight on my part and explain the reason for this failing of my code. however if it does not come to pass that I find the error I will at least gain experience in using git hub's compare branches function and also I will redo the same section again but this time with the instructor's example beginning code. as perhaps there may have been some kind of subtle deviation that i introduced long ago in the guided walkthrough that made my organization too radically different to work as showed however given my understanding of copnent binding and communication i just cannot see where in the chain I went wrong even if my file structure and architecture does differ there are no syntax errors to be found.